Make ICE connections wire equivalent with plain UDP.
(Once NAT is pierced, and assuming we are not relaying through TURN, and
ignoring the actual STUN packets.)

Refactored code so that almost all of the UDP code is shared.

Having this exact wire compatibility is really important for the future.
I had assumed that the WebRTC code was putting some connection IDs, etc
in the envelope, but I was wrong.  I believe that is happening at a higher
layer in their code (the DTLS layer), and I'm below that.

I am pretty sur ethat I could get rid of the special "ping check" messages,
and in fact they probably are just delaying the connection and I should
trust the WebRTC code.  Will probably revisit this later.