Fix bug in expiry of in-flight packets.
This meant the map was growing without bound!  The worse the packet
loss, the worse the growth of this map.  This didn't break the wire
protocol, it was just bad for perf.  But not that bad, because the
data structure has a good O() time and all of the old packets that
should have been expired were just not really examined.

Until I added code to check it constantly.  Then I noticed perf has
tanked!

I am actually surprised that static analysis didn't catch this.