Several P2P improvements
Add a connection time to CMsgSteamDatagramLinkLifetimeStats.
Connection will track when it entered and left theconnected
state, so that when it fills out this record, it can populate
this field.

Change default value of P2P_Transport_ICE_Enable.  We don't have a
"platform" to ask for user preference in this opernsource lib.

Adjust some asserts so that if they fire because some transport
is trying to process packets in the wrong state, we won't crash.

TRack how long each P2P transport was selected.

Added some SDR transport session stats (in Steam).

Fix a compiler warning.

Remove an instance where RTTI is used.  (Many people prefer to
compile without RTTI.)