Fix inverted logic
Revert part of 705bc0e4bbbb891cf797dfb8e5595efaefeb1e31, which managed to
refactor code without improving it in any way AND introduce an inverted
logic bug.

This causes us to spam panic keepalives when we are timing out much
faster than intended.

Also, when deciding if it is time to take action, use inclusive comparisons.
If we ask to wake up at a certain time, and we do actually manage to wake up
at that exact time, then we should take action.  This is probably extremely
rare, however in an environment with limited timer precision, it could be
more common.