Tweak handling of some edge cases.
In several different edge cases, we may decide not to ack a packet to the peer.
(Especially if nwe need to protect ourselves against a malicious peer, etc.)

In that case, don't drop the whole packet and stop processing it.  That is too
much.  Just inhibit marking the packet as received and acking it.  Among other
things, we want our stats to reflect the fact that we did receive this packet.
Especially, we want to mark the packet as received, so that we can quickly
reject a duplicate.

Also add a bunch of clarifying comments.