Fix bug classifying P2P failure reason
Just because we did not gather any reflexive address that was enabled, that
we failed to STUN.

"Failed to gather any" is ambiguous English.  We really meant "we did not gather
any".

Also, if we have a public IPv4 address, it could be that we did not STUN,
because the STUN server told us that our address was already public, and so
was not gathered as an additional address.