Use variadic AssertMsg macro
Bring it over from Steam branch.  This was causing tons of little nagging
differences that constantly had to be fixed, since I would use AssertMsg()
in Steam, and then it wouldn't compeil here, and I had to change to
AssertMsg1(), etc.

This uses some fancy template magic so that if there are no arguments,
it does not use printf